+++ js/filter.js
@@ -1,0 +1,1 @@
+//Модуль фильтрации обьявлений
