+++ js/map.js
@@ -1,0 +1,1 @@
+//Модуль для загрузки карты
