+++ js/submit-form.js
@@ -1,0 +1,1 @@
+//Модуль для обработки отправки формы
